Display ToDo index
Users need to know the index of the ToDO item to successfully
mark it as complete. Without the index, it is just a guess as to
what the index might be. Completed items are not displayed so that
makes it difficult to know the index of an uncompleted ToDo without
viewing the saved JSON.